Update node.h
I set up a basic class to represent each node/spot on the grid and any markings it may have. I included some simple functions like the constructor, getters, and a few bools to keep track of what is around it. I think this will help us organize everything we need to keep track of :)